+++ PestrikovBank/Application/SceenDelegate.swift
@@ -10,6 +10,8 @@ import UIKit
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
     var window: UIWindow?
+    var appCoordinator: RouteCoordinator?
+    var appDIContainer: AppDIContainer = AppDIContainer()
 
     func scene(
         _ scene: UIScene,
@@ -19,14 +21,13 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         self.window = UIWindow(windowScene: windowScene)
         
-        let model = LoginViewModelImpl()
-        window?.rootViewController = LoginViewControllerImpl(
-            loginForm: LoginFormViewImpl(
-                frame: window?.bounds ?? UIScreen.main.bounds,
-                phoneNumberFormatting: model.phoneFormat()
-            ),
-            viewModel: model
-        )
+        let appCoordinator = appDIContainer
+            .routeCoordinatorDIContainer
+            .appCooridnator
+        appCoordinator.start()
+        
+        self.appCoordinator = appCoordinator
+        self.window?.rootViewController = appCoordinator.rootViewController
         window?.makeKeyAndVisible()
     }
 }
