+++ PestrikovBank/Domain/Error/PBError.swift
@@ -10,8 +10,9 @@ import Foundation
 public enum PBError: LocalizedError {
     case invalidResponse
     case authError(String)
-    case networkError
-    case serverError
+    case network(Error)
+    case keychain(Error)
+    case server(statusCode: Int)
     
     public var errorDescription: String? {
         switch self {
@@ -19,10 +20,12 @@ public enum PBError: LocalizedError {
             return "Invalid response format, try again later or contact support"
         case .authError(let message):
             return "Authentication error: \(message)"
-        case .networkError:
-            return "No internet connection"
-        case .serverError:
-            return "Ooops! Something went wrong"
+        case .network(let e):
+            return "No internet connection. \(e.localizedDescription)"
+        case .server(let e):
+            return "Ooops! Something went wrong. Server error: \(e)"
+        case .keychain(_):
+            return "Can not fetch keychain data"
         }
     }
 }
