+++ PestrikovBank/Domain/UseCases/Auth/LoginUseCaseImpl.swift
@@ -1,0 +1,14 @@
+//
+//  LoginUseCaseImpl.swift
+//  PestrikovBank
+//
+//  Created by m on 19.04.2025.
+//
+
+import Combine
+
+//public struct LoginUseCaseImpl: LoginUseCase {
+//    public func execute(with creds: any Creds) -> AnyPublisher<any AuthTokens, any Error> {
+//        <#code#>
+//    }
+//}
