--- PestrikovBank/Presentation/Protocols/Interceptor/PhoneFormat.swift
@@ -1,15 +1,0 @@
-//
-//  PhoneFormat.swift
-//  PestrikovBank
-//
-//  Created by m on 19.04.2025.
-//
-
-import Foundation
-
-public protocol PhoneFormat {
-    var maxLength: Int { get }
-    var regex: NSRegularExpression { get }
-
-    func phoneFormat() -> (String, Bool) -> String
-}
