+++ PestrikovBank/Presentation/Routing/MainCoordinator.swift
@@ -1,0 +1,18 @@
+//
+//  MainCoordinator.swift
+//  PestrikovBank
+//
+//  Created by m on 20.04.2025.
+//
+
+import UIKit
+
+public class MainCoordinator: RouteCoordinator {
+    
+    public var rootViewController: UIViewController = UITabBarController()
+    // var childCoordinators = [RouteCoordinator]()
+    
+    public func start() {
+        // todo
+    }
+}
