+++ PestrikovBank/Resorces/Assets.xcassets/LoginLightBackground.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "gradient.jpeg",
+      "filename" : "meshGradient (1).png",
       "idiom" : "universal",
       "scale" : "1x"
     },
